# Fix StringSegment benchmarks