@danmoseley 2021-02-12
@billwert  are you an admin on this repo? If so you probably want to disable merge and rebase commits. I almost did a merge commit. The runtime repo requires squashes.