@billwert 2021-02-12
Good tip! Will do so

Thanks,
-Bill
________________________________
From: Dan Moseley <notifications@github.com>
Sent: Thursday, February 11, 2021 6:25:18 PM
To: dotnet/performance <performance@noreply.github.com>
Cc: Bill Wert <billwert@microsoft.com>; Mention <mention@noreply.github.com>
Subject: Re: [dotnet/performance] Fix StringSegment benchmarks (#1661)


@billwert<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fbillwert&data=04%7C01%7Cbillwert%40microsoft.com%7C62d76917db6f4d11d35108d8cefd70ad%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637486935211355061%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C1000&sdata=WY1Zr6l1fphBkKIURbCQ%2BVQxTbMTNNLiyWtBLhT%2BYc4%3D&reserved=0> are you an admin on this repo? If so you probably want to disable merge and rebase commits. I almost did a merge commit. The runtime repo requires squashes.

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fdotnet%2Fperformance%2Fpull%2F1661%23issuecomment-777926675&data=04%7C01%7Cbillwert%40microsoft.com%7C62d76917db6f4d11d35108d8cefd70ad%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637486935211365060%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C1000&sdata=9JXhTD2orQ8B68X3ZQqHblSltoQ%2Bv1gSqyI1Vm7SGiY%3D&reserved=0>, or unsubscribe<https://nam06.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FABCQI4FNZLHTIEOQXTNFMCTS6SGQ5ANCNFSM4XP2CUJA&data=04%7C01%7Cbillwert%40microsoft.com%7C62d76917db6f4d11d35108d8cefd70ad%7C72f988bf86f141af91ab2d7cd011db47%7C1%7C0%7C637486935211365060%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C1000&sdata=jdWRLF0qhJ5xHS3AJlDWHfmKH8uyz4MO0NZ7CahymRs%3D&reserved=0>.